feat: use cobra to handle cmd
